requirements.txt
Django==1.6.5
dj-database-url==0.3.0
dj-static==0.0.6
gunicorn==19.1.1
psycopg2==2.5.1
static==0.4
wsgiref==0.1.2
South==1.0
sendgrid==1.5.20
image==1.4.4
Pillow==3.1.0